requirements.txt
Flask>=3.1.0
flaskwebgui>=1.1.0
waitress>=3.0.0
psutil>=7.0.0
pillow>=10.0.0
screeninfo>=0.8.1

